requirements.txt
pandas
numpy
scikit-learn
scipy
tqdm
pyyaml
orjson
joblib
statsmodels
matplotlib
networkx

# Dev / QA tools
ruff
black
pytest

openai>=1.30.0
dashscope>=1.19.0


